+++ packages/manager-components/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.24.0](https://github.com/ovh/manager/compare/@ovhcloud/manager-components@1.23.0...@ovhcloud/manager-components@1.24.0) (2024-08-01)
+
+### Features
+
+- pci databases analytics app ([#10914](https://github.com/ovh/manager/issues/10914)) ([3d589ef](https://github.com/ovh/manager/commit/3d589efe6f60d3ea18d555a838038ed7ca5c9432))
+
 # [1.23.0](https://github.com/ovh/manager/compare/@ovhcloud/manager-components@1.22.1...@ovhcloud/manager-components@1.23.0) (2024-07-18)
 
 ### Bug Fixes