+++ packages/manager/apps/pci/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.6.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.6.5...@ovh-ux/manager-pci-app@4.6.6) (2024-08-01)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
 ## [4.6.5](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.6.4...@ovh-ux/manager-pci-app@4.6.5) (2024-07-26)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-app