+++ packages/manager/apps/vrack-services/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-vrack-services-app@0.5.3...@ovh-ux/manager-vrack-services-app@0.5.4) (2024-08-01)
+
+**Note:** Version bump only for package @ovh-ux/manager-vrack-services-app
+
+
+
+
+
 ## [0.5.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-vrack-services-app@0.5.2...@ovh-ux/manager-vrack-services-app@0.5.3) (2024-07-18)
 
 **Note:** Version bump only for package @ovh-ux/manager-vrack-services-app