+++ packages/manager/modules/pci/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.56.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci@6.55.2...@ovh-ux/manager-pci@6.56.0) (2024-08-01)
+
+
+### Features
+
+* pci databases analytics app ([#10914](https://github.com/ovh/manager/issues/10914)) ([3d589ef](https://github.com/ovh/manager/commit/3d589efe6f60d3ea18d555a838038ed7ca5c9432))
+
+
+
+
+
 ## [6.55.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci@6.55.1...@ovh-ux/manager-pci@6.55.2) (2024-07-26)